+++ catalog/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Category, Product, ContactData
+from .models import Category, Product
 
 @admin.register(Category)
 class CategoryAdmin(admin.ModelAdmin):
@@ -10,7 +10,3 @@ class ProductAdmin(admin.ModelAdmin):
     list_display = ('id', 'name', 'price', 'category')
     list_filter = ('category',)
     search_fields = ('name', 'description')
-
-@admin.register(ContactData)
-class ContactDataAdmin(admin.ModelAdmin):
-    list_display = ('name', 'value')
