+++ catalog/urls.py
@@ -3,5 +3,4 @@ from catalog import views
 
 urlpatterns = [
     path('', views.home, name='home'),
-    path('contacts/', views.contacts_view, name='contacts'),
 ]
