+++ contacts/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+from .models import ContactData
+
+@admin.register(ContactData)
+class ContactDataAdmin(admin.ModelAdmin):
+    list_display = ('name', 'value')
+
