+++ contacts/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+class ContactsConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "contacts"
+    verbose_name = "Обратная связь"
