+++ contacts/forms.py
@@ -1,0 +1,6 @@
+from django import forms
+
+class ContactForm(forms.Form):
+    name = forms.CharField(label="Имя", max_length=100)
+    email = forms.EmailField(label="Email")
+    message = forms.CharField(label="Сообщение", widget=forms.Textarea)
