+++ templates/catalog/base.html
@@ -13,7 +13,7 @@
     <div>
       <ul class="navbar-nav">
         <li class="nav-item"><a class="nav-link" href="/">Главная</a></li>
-        <li class="nav-item"><a class="nav-link" href="/contacts/">Контакты</a></li>
+        <li class="nav-item"><a class="nav-link" href="/contacts">Контакты</a></li>
       </ul>
     </div>
   </div>
